[FEAT]: Breaks Pill WIdgets down into separate files
