+++ lib/widgets/molecules/pills/gt_pill.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:gt_mobile_foundation/foundation.dart';
 import 'package:gt_mobile_ui/gt_mobile_ui.dart';
 
-/// Defines the visual variants available for status and button pills.
+/// Defines the visual variants available for status, button, and other pills.
+/// Each variant determines the background, border, and text colors applied to the pill.
 enum GtPillVariant {
-  /// A strong emphasis variant, typically using solid, bold colors.
+  /// A strong emphasis variant, typically using solid, bold brand colors.
   strong,
 
   /// A neutral variant, typically using subdued gray or muted tones.
@@ -42,9 +42,9 @@ enum GtPillVariant {
   primary,
 }
 
-/// Internal extension to compute color schemes for pill variants.
-extension on GtPillVariant? {
-  /// Returns the appropriate text color for the pill variant based on the provided [palette].
+/// Internal extension to compute color schemes for [GtPillVariant] elements.
+extension GtPillExtension on GtPillVariant? {
+  /// Returns the appropriate text color for this pill variant based on the provided [palette].
   Color getTextColor(GtPalette palette) {
     return switch (this) {
       .primary => palette.primary.dark,
@@ -62,7 +62,7 @@ extension on GtPillVariant? {
     };
   }
 
-  /// Returns the appropriate border color for the pill variant based on the provided [palette].
+  /// Returns the appropriate border color for this pill variant based on the provided [palette].
   Color getBorderColor(GtPalette palette) {
     return switch (this) {
       .primary => palette.primary.alpha16,
@@ -80,7 +80,7 @@ extension on GtPillVariant? {
     };
   }
 
-  /// Returns the appropriate background color for the pill variant based on the provided [palette].
+  /// Returns the appropriate background color for this pill variant based on the provided [palette].
   Color getBgColor(GtPalette palette) {
     return switch (this) {
       .primary => palette.primary.alpha10,
@@ -99,9 +99,9 @@ extension on GtPillVariant? {
   }
 }
 
-/// Defines the available sizes for pill widgets.
+/// Defines the available sizes for pill widgets, influencing padding and overall scale.
 enum GtPillSize {
-  /// The standard, default pill size.
+  /// The standard, default pill size used in most standard contexts.
   normal,
 
   /// A slightly larger pill size for better visibility or touch area.
@@ -109,8 +109,10 @@ enum GtPillSize {
 }
 
 /// A foundational pill widget that displays text and an optional icon within a rounded container.
+///
+/// This serves as the base structure for other specialized pills (e.g., [GtStatusPill], [GtButtonPill]).
 class GtPill extends StatelessWidget {
-  /// The text to display inside the pill. It will be automatically converted to uppercase.
+  /// The text to display inside the pill. In some implementations, this may be automatically converted to uppercase.
   final String text;
 
   /// An optional icon widget to display alongside the text.
@@ -119,25 +121,25 @@ class GtPill extends StatelessWidget {
   /// An optional icon widget to display after the text.
   final Widget? trailing;
 
-  /// The internal padding of the pill. Defaults to 4.px on all sides.
+  /// The internal padding of the pill. Defaults to 4 pixels on all sides if not provided.
   final EdgeInsets? padding;
 
-  /// The background color of the pill.
+  /// The fill or background color of the pill's container.
   final Color bgColor;
 
-  /// The optional border color. If null, the border defaults to [bgColor].
+  /// The border color. If null, the border color defaults to the [bgColor], effectively hiding the border.
   final Color? borderColor;
 
-  /// The color of the text and potentially the icon (if it inherits the text color).
+  /// The color of the text. Icons will inherit this color unless overridden.
   final Color? textColor;
 
-  /// The alignment of the child within the pill's container.
+  /// The alignment of the content within the pill's internal bounds.
   final Alignment? alignment;
 
-  /// Whether to display a drop shadow beneath the pill.
+  /// Determines if a drop shadow is rendered beneath the pill container.
   final bool showShadow;
 
-  /// The visual variant of the pill, determining its contextual colors.
+  /// The visual variant that informs context-based color calculations.
   final GtPillVariant variant;
 
   /// The custom border radius of the pill. Defaults to the theme's small border radius.
@@ -208,435 +210,5 @@ class GtPill extends StatelessWidget {
     }
 
     return child;
-  }
-}
-
-/// A specialized pill widget used to indicate status, featuring predefined color variants and a compact layout.
-class GtStatusPill extends StatelessWidget {
-  /// The status text to display.
-  final String text;
-
-  /// The visual variant determining the colors of the pill.
-  final GtPillVariant? variant;
-
-  /// An optional icon data to display. Rendered at size 11.
-  final IconData? icon;
-
-  /// An optional icon widget to display after the text.
-  final IconData? trailing;
-
-  /// The alignment of the content within the pill.
-  final Alignment? alignment;
-
-  /// The size of the status pill.
-  final GtPillSize size;
-
-  /// Creates a [GtStatusPill].
-  const GtStatusPill({
-    super.key,
-    required this.text,
-    this.variant,
-    this.trailing,
-    this.icon,
-    this.alignment,
-    this.size = .normal,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final palette = context.palette;
-    final textColor = variant.getTextColor(palette);
-    final bgColor = variant.getBgColor(palette);
-    final borderColor = variant.getBorderColor(palette);
-    Widget? iconWidget;
-    Widget? trailingWidget;
-
-    if (icon != null) {
-      iconWidget = GtIcon.withColor(icon!, color: textColor, size: 11);
-    }
-
-    if (trailing != null) {
-      trailingWidget = GtIcon.withColor(trailing!, color: textColor, size: 11);
-    }
-
-    final padding = switch (size) {
-      .larger => 6.px,
-      _ => 4.px,
-    };
-
-    return GtPill(
-      text: text.upper,
-      bgColor: bgColor,
-      icon: iconWidget,
-      trailing: trailingWidget,
-      textColor: textColor,
-      borderColor: borderColor,
-      padding: context.insets.allDp(padding),
-      alignment: alignment,
-      variant: variant ?? .strong,
-    );
-  }
-}
-
-/// A specialized pill widget intended for tags or interactive-looking badges.
-///
-/// Features slightly larger padding and a larger icon size compared to [GtStatusPill].
-class GtButtonPill extends StatelessWidget {
-  /// The text to display.
-  final String text;
-
-  /// The visual variant determining the colors of the pill.
-  final GtPillVariant? variant;
-
-  /// An optional icon data to display. Rendered at size 14.
-  final IconData? icon;
-
-  /// An optional icon widget to display after the text.
-  final IconData? trailing;
-
-  /// The alignment of the content within the pill.
-  final Alignment? alignment;
-
-  /// An optional callback triggered when the pill is tapped.
-  final OnPressed? onTap;
-
-  /// Whether to display a drop shadow beneath the pill.
-  final bool showShadow;
-
-  /// The size of the button pill.
-  final GtPillSize size;
-
-  /// Creates a [GtButtonPill].
-  const GtButtonPill({
-    super.key,
-    required this.text,
-    this.variant,
-    this.icon,
-    this.trailing,
-    this.alignment,
-    this.onTap,
-    this.showShadow = false,
-    this.size = .normal,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final palette = context.palette;
-    final textColor = variant.getTextColor(palette);
-    final bgColor = variant.getBgColor(palette);
-
-    Widget? iconWidget;
-    Widget? trailingWidget;
-
-    if (icon != null) {
-      iconWidget = GtIcon.withColor(icon!, color: textColor, size: 14);
-    }
-
-    if (trailing != null) {
-      trailingWidget = GtIcon.withColor(trailing!, color: textColor, size: 14);
-    }
-
-    final padding = switch (size) {
-      .larger => 8.px,
-      _ => 6.px,
-    };
-
-    return InkWell(
-      onTap: () {
-        if (onTap == null) return;
-        HapticFeedback.lightImpact();
-        onTap?.call();
-      },
-      child: GtPill(
-        text: text.upper,
-        bgColor: bgColor,
-        borderColor: bgColor,
-        icon: iconWidget,
-        textColor: textColor,
-        padding: context.insets.allDp(padding),
-        trailing: trailingWidget,
-        alignment: alignment,
-        showShadow: showShadow,
-        variant: variant ?? .strong,
-      ),
-    );
-  }
-}
-
-/// A specialized pill widget intended for displaying informational text and badges.
-///
-/// Features a fully rounded, capsule-like shape and supports custom typography scaling.
-class GtInfoPill extends StatelessWidget {
-  /// The text to display.
-  final String text;
-
-  /// The visual variant determining the colors of the pill.
-  final GtPillVariant? variant;
-
-  /// An optional icon data to display. Rendered at size 14.
-  final IconData? icon;
-
-  /// An optional icon widget to display after the text.
-  final IconData? trailing;
-
-  /// The alignment of the content within the pill.
-  final Alignment? alignment;
-
-  /// Whether to display a drop shadow beneath the pill.
-  final bool showShadow;
-
-  /// Whether to use the display typography style instead of standard body text.
-  final bool useDisplayFont;
-
-  /// Creates a [GtInfoPill].
-  const GtInfoPill({
-    super.key,
-    required this.text,
-    this.variant,
-    this.icon,
-    this.trailing,
-    this.alignment,
-    this.showShadow = false,
-    this.useDisplayFont = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final palette = context.palette;
-    final textColor = variant.getTextColor(palette);
-    final bgColor = variant.getBgColor(palette);
-    final verticalPadding = useDisplayFont ? 4.px : 2.px;
-    TextStyle style = context.textStyles.bodyS(color: textColor);
-
-    if (useDisplayFont) {
-      style = context.textStyles.titleXs(color: textColor);
-    }
-
-    Widget? iconWidget;
-    Widget? trailingWidget;
-
-    if (icon != null) {
-      iconWidget = GtIcon.withColor(icon!, color: textColor, size: 18);
-    }
-
-    if (trailing != null) {
-      trailingWidget = GtIcon.withColor(trailing!, color: textColor, size: 18);
-    }
-
-    return GtPill(
-      text: text,
-      bgColor: bgColor,
-      borderColor: bgColor,
-      icon: iconWidget,
-      borderRadius: context.borderRadius4Xl,
-      textStyle: style,
-      padding: context.insets.symmetricDp(
-        vertical: verticalPadding,
-        horizontal: 8.px,
-      ),
-      trailing: trailingWidget,
-      alignment: alignment,
-      showShadow: showShadow,
-      variant: variant ?? .strong,
-    );
-  }
-}
-
-/// A specialized pill widget intended for tabbed selections.
-///
-/// Automatically handles selected and unselected states by comparing the [value]
-/// against the [activeValue].
-class GtTabPill<T> extends StatelessWidget {
-  /// The underlying value represented by this tab.
-  final T value;
-
-  /// The currently active/selected value in the tab group.
-  final T? activeValue;
-
-  /// The text to display.
-  final String text;
-
-  /// The visual variant determining the colors of the pill.
-  final GtPillVariant? variant;
-
-  /// An optional icon data to display. Rendered at size 14.
-  final IconData? icon;
-
-  /// An optional icon widget to display after the text.
-  final IconData? trailing;
-
-  /// The alignment of the content within the pill.
-  final Alignment? alignment;
-
-  /// The callback triggered when the tab is selected.
-  final OnChanged<T> onSelect;
-
-  /// Creates a [GtTabPill].
-  const GtTabPill({
-    super.key,
-    required this.text,
-    required this.value,
-    required this.activeValue,
-    required this.onSelect,
-    this.variant,
-    this.icon,
-    this.trailing,
-    this.alignment = .centerLeft,
-  });
-
-  /// Creates a specialized selection variant of the tab pill.
-  const factory GtTabPill.selection({
-    Key? key,
-    required String text,
-    required T value,
-    required T? activeValue,
-    required OnChanged<T> onSelect,
-    GtPillVariant variant,
-    IconData? icon,
-    IconData? trailing,
-    Alignment? alignment,
-  }) = GtSelectionPill<T>;
-
-  /// Gets the appropriate text color based on the selection state.
-  Color getTextColor(GtPalette palette) {
-    if (!isSelected) return variant.getTextColor(palette);
-    return switch (variant) {
-      .neutral => palette.text.darkerSub,
-      _ => palette.text.white,
-    };
-  }
-
-  /// Gets the appropriate background color based on the selection state.
-  Color getBgColor(GtPalette palette) {
-    if (!isSelected) return variant.getBgColor(palette);
-    return switch (variant) {
-      .primary => palette.primary.base,
-      .neutral => palette.bg.sub,
-      .featured => palette.feature.base,
-      .info => palette.information.base,
-      .success => palette.success.base,
-      .warning => palette.warning.base,
-      .error => palette.error.base,
-      .highlighted => palette.highlighted.base,
-      .stable => palette.stable.base,
-      .verified => palette.success.base,
-      .away => palette.away.base,
-      _ => palette.bg.strong,
-    };
-  }
-
-  /// Automatically scrolls this tab into view if it is selected.
-  void scrollIntoView(BuildContext context) {
-    if (!isSelected) return;
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      Scrollable.ensureVisible(context);
-    });
-  }
-
-  /// Whether this tab is currently the active selected tab.
-  bool get isSelected => activeValue == value;
-
-  @override
-  Widget build(BuildContext context) {
-    scrollIntoView(context);
-    final palette = context.palette;
-    final textColor = getTextColor(palette);
-    final bgColor = getBgColor(palette);
-    Widget? iconWidget;
-    Widget? trailingWidget;
-
-    if (icon != null) {
-      iconWidget = GtIcon.withColor(icon!, color: textColor, size: 16);
-    }
-
-    if (trailing != null) {
-      trailingWidget = GtIcon.withColor(trailing!, color: textColor, size: 16);
-    }
-
-    return InkWell(
-      onTap: () {
-        HapticFeedback.selectionClick();
-        onSelect(value);
-      },
-      child: GtPill(
-        text: text.upper,
-        bgColor: bgColor,
-        borderColor: bgColor,
-        textStyle: context.textStyles.button2s(color: textColor),
-        icon: iconWidget,
-        padding: context.insets.symmetricDp(vertical: 8.px, horizontal: 12.px),
-        trailing: trailingWidget,
-        alignment: alignment,
-        variant: variant ?? .strong,
-      ),
-    );
-  }
-}
-
-/// A specialized selection pill widget intended for inline selectable items.
-///
-/// Similar to [GtTabPill], but features distinct unselected styling with a
-/// transparent background and modified typography.
-class GtSelectionPill<T> extends GtTabPill<T> {
-  /// Creates a [GtSelectionPill].
-  const GtSelectionPill({
-    super.key,
-    required super.text,
-    required super.value,
-    required super.activeValue,
-    required super.onSelect,
-    super.variant,
-    super.icon,
-    super.trailing,
-    super.alignment = .centerLeft,
-  });
-
-  @override
-  Color getTextColor(GtPalette palette) {
-    if (!isSelected) return palette.text.darkerSub;
-    return super.getTextColor(palette);
-  }
-
-  @override
-  Color getBgColor(GtPalette palette) {
-    if (!isSelected) return Colors.transparent;
-    return super.getBgColor(palette);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    scrollIntoView(context);
-    final palette = context.palette;
-    final textColor = getTextColor(palette);
-    final bgColor = getBgColor(palette);
-    Widget? iconWidget;
-    Widget? trailingWidget;
-
-    if (icon != null) {
-      iconWidget = GtIcon.withColor(icon!, color: textColor, size: 16);
-    }
-
-    if (trailing != null) {
-      trailingWidget = GtIcon.withColor(trailing!, color: textColor, size: 16);
-    }
-
-    return InkWell(
-      onTap: () {
-        HapticFeedback.selectionClick();
-        onSelect(value);
-      },
-      child: GtPill(
-        text: text.capitalise(),
-        bgColor: bgColor,
-        borderColor: bgColor,
-        borderRadius: context.borderRadiusMd,
-        textStyle: context.textStyles.subHeadS(color: textColor),
-        icon: iconWidget,
-        padding: context.insets.symmetricDp(vertical: 4.px, horizontal: 12.px),
-        trailing: trailingWidget,
-        alignment: alignment,
-        variant: variant ?? .strong,
-      ),
-    );
   }
 }
