+++ lib/widgets/molecules/pills/pills.dart
@@ -1,1 +1,5 @@
+export 'gt_button_pill.dart';
+export 'gt_info_pill.dart';
 export 'gt_pill.dart';
+export 'gt_status_pill.dart';
+export 'gt_tab_pill.dart';
